move ModinIndex class; more docs
